fix(svo,generation): Added Cell::get_path to fix follow_path not going inside packed cells

+++ svo/src/packed.rs
@@ -15,7 +15,7 @@ impl<D> PackedCellLevel<D> {
     {
         Self {
             data: {
-                let size = 8u32.pow(depth) as usize;
+                let size = 8usize.pow(depth);
                 let mut vec = Vec::with_capacity(size);
                 for _ in 0..size.saturating_sub(1) {
                     vec.push(data.clone());
